api/requirements.txt
pymssql
pandas
azure-functions
